Fix type

+++ R/git.R
@@ -124,7 +124,7 @@ git_commits <- function(branch, root = NULL) {
   commits <- git_run(args, root = root, check = TRUE)$output
   commits <- utils::read.table(text = commits, stringsAsFactors = FALSE,
                                sep = ",", col.names = c("id", "date_time"),
-                               colClasses = c("character", "numeric"))
+                               colClasses = c("character", "integer"))
   commits$age <- calculate_age(commits$date_time)
   commits$date_time <- convert_unix_to_iso_time(commits$date_time)
   commits
